add try catch block

+++ Back end/New folder (3)/bloomevents/src/main/java/com/project/bloomevents/ServiceImpl/ReviewServiceImpl.java
@@ -1,5 +1,6 @@
 package com.project.bloomevents.ServiceImpl;
 
+import com.project.bloomevents.CommonResponse.CommonResponse;
 import com.project.bloomevents.DTO.ReviewDTO;
 import com.project.bloomevents.Model.Review;
 import com.project.bloomevents.Repository.ReviewRepository;
@@ -18,24 +19,63 @@ public class ReviewServiceImpl implements ReviewService {
     @Autowired
     private ModelMapper modelMapper;
 
+    public ReviewDTO getReviewById(int reviewId){
+        try{
+            Review review = reviewRepo.getReviewById(reviewId);
+            if (review == null) {
+                return null;
+            }
+            return modelMapper.map(review, new TypeToken<ReviewDTO>() {
+            }.getType());
+        }
+        catch(Exception e){
+            System.out.println(e.toString());
+            return null;
+        }
+    }
+
     @Override
     public List<ReviewDTO> getAllReviews() {
-        List<Review> list= reviewRepo.findAll();
-        return modelMapper.map(list, new TypeToken<List<ReviewDTO>>(){}.getType());
+        try{
+            List<Review> list = reviewRepo.findAll();
+            return modelMapper.map(list, new TypeToken<List<ReviewDTO>>() {
+            }.getType());
+        }
+        catch(Exception e){
+            System.out.println(e.toString());
+            return null;
+        }
     }
 
     @Override
     public ReviewDTO addReview(ReviewDTO review) {
-        Review r= modelMapper.map(review,Review.class);
-        Review r1=reviewRepo.save(r);
+        try{
+            Review r = modelMapper.map(review, Review.class);
+            Review r1 = reviewRepo.save(r);
 
-        return modelMapper.map(r1,new TypeToken<ReviewDTO>(){}.getType());
+            return modelMapper.map(r1, new TypeToken<ReviewDTO>() {
+            }.getType());
+        }
+        catch(Exception e){
+            System.out.println(e.toString());
+            return null;
+        }
     }
 
     @Override
-    public String deleteReview(int reviewId) {
-        reviewRepo.deleteById(reviewId);
-        return "review deleted";
-
+    public CommonResponse deleteReview(int reviewId) {
+        try{
+            ReviewDTO review = getReviewById(reviewId);
+            if (review != null) {
+                reviewRepo.deleteById(reviewId);
+                return new CommonResponse(true, "Successfully deleted the review :" + reviewId);
+            } else {
+                return new CommonResponse(false, "Cannot find a review with id :" + reviewId);
+            }
+        }
+        catch(Exception e){
+            System.out.println(e.toString());
+            return null;
+        }
     }
 }
